+++ src/app/components/VideoCard.tsx
@@ -1,4 +1,4 @@
-// app/components/VideoItem.tsx
+// src/app/components/VideoItem.tsx
 import React, { useRef, useState, useEffect } from 'react';
 import VideoItem from '../lib/definitions/VideoItem';
 
@@ -21,9 +21,9 @@ export default function VideoCard({ item, onClick }: VideoItemProps) {
     }
   }, []);
 
-  useEffect(() => {
-    console.log('VideoItem received:', item);
-  });
+  // useEffect(() => {
+  //   console.log('VideoItem received:', item);
+  // });
 
   const handleVideoLoad = () => {
     setIsLoaded(true);